For spacecookie >= 0.3.0.0 prefixing with 'i' is no longer needed
The changes are described in detail in the changelog:
https://github.com/sternenseemann/spacecookie/blob/master/CHANGELOG.md#changes-to-networkgopherutilgophermap

In a nutshell we now require a gophermap line which links to another
gopher resource to contain at least one tab character. This makes them
easily distinguishable from info lines and resolves the issue where an
info line would be parsed incorrectly if it started with a valid gopher
file type character.

Unfortunately this also breaks the previous workaround for the issue
which was used in burrow:

  iMy text bla bla

This now can be written as

  My text bla bla

or

  iMy text bla bla\t

to achieve the same result.

This commit adjusts burrow's code for the first variant and updates a
few documentation links.