explicit about example + more reliable building
It's been a while since I've developed this
project. I come back and it's hard to compile!

So I got it working in Nix and regular Cabal,
hopefully future-proofing it.

I've included two different cabal freezes,
one for nix and one for not-nix. I may try to
figure out how to just have one cabal freeze
for both.